+++ einguteswerkzeug/shell/__init__.py
@@ -4,8 +4,11 @@
 """
 import cmd
 import logging
+import json
+import pydoc
 import string
 from pluginbase import PluginBase
+from einguteswerkzeug.helpers import editor, confirm_prompt
 from einguteswerkzeug.plugins import EGWPluginFilter, EGWPluginGenerator
 from einguteswerkzeug.core.template import EGWTemplate, load_templates, select_template
 from einguteswerkzeug.core import EGW
@@ -26,7 +29,7 @@ log.addHandler(handler)
 log.propagate=False
 # ---
 
-__version__ = (0,0,1)
+__version__ = (0,1,0)
 
 
 class EGWShell(cmd.Cmd):
@@ -35,9 +38,18 @@ class EGWShell(cmd.Cmd):
     einguteswergzeug (egw).
     """
     __version__ = __version__
-    intro = 'Welcome to the egw-shell shell. Type help or ? to list commands.\n'
+    intro = 'Welcome to egw-shell ({}). Type help or ? to list commands.\n'.format('.'.join(str(i) for i in __version__))
     prompt = '(egw) '
     egw = None # egw instance
+    # ---colors ... MAGENTA YELLOW LIGHTBLUE_EX LIGHTGREEN_EX CYAN LIGHTCYAN_EX
+    # TODOP3 move to config-file
+    #color_prompt = Fore.LIGHTGREEN_EX
+    color_prompt = Fore.GREEN
+    color_gen  = Fore.LIGHTCYAN_EX
+    color_gen2 = Fore.CYAN
+    color_flt  = Fore.LIGHTYELLOW_EX
+    color_flt2 = Fore.YELLOW
+    # ---
 
     def __init__(self, *args, **kwargs):
         print(kwargs)
@@ -47,25 +59,42 @@ class EGWShell(cmd.Cmd):
     def do_EOF(self, line):
         return True
 
-    def postloop(self):
-        print
+    def preloop(self):
+        print(self.color_prompt)
 
-    def do_listGenerators(self,arg):
+    def postloop(self):
+        print(Style.RESET_ALL)
+
+    def postcmd(self, stop, line):
+        print(self.color_prompt)
+
+    def do_listGenerators(self):
         """
         list available generators
         """
-        #print(Fore.CYAN) #MAGENTA YELLOW LIGHTBLUE_EX LIGHTGREEN_EX CYAN LIGHTCYAN_EX
         for i, g in enumerate(self.egw.generators.values()):
-            print(Fore.LIGHTBLUE_EX + "{} {} - {}".format(i, g.name, g.description) + Style.RESET_ALL)
+            print(self.color_gen + "{} {} - {}".format(i, g.name, g.description) + Style.RESET_ALL)
+
+    def do_generators(self, arg):
+        """
+        list available generators
+        """
+        self.do_listGenerators()
 
     def do_listFilters(self,arg):
         """
         list available filters
         """
-        print(Fore.CYAN)
+        print(self.color_flt)
         for i, f in enumerate(self.egw.filters.values()):
             print("{} {} - {}".format(i, f.name, f.description))
         print(Style.RESET_ALL)
+
+    def do_filters(self, arg):
+        """
+        list available filters
+        """
+        self.do_listFilters(None)
 
     def do_detailGenerator(self, generator):
         """detailGenerator [generator]
@@ -76,27 +105,30 @@ class EGWShell(cmd.Cmd):
         else:
             g = self.egw.generators[generator]
 
-
         header  = "{}".format(g.name)
+        header  += "\n" +"="*len(g.name)
         tpl = string.Template("""
-        help
-        ----
-        $help
+help
+----
+$help
+
+parameters ('--params-generator')
+---------------------------------
+$kwargs
         """)
-
         content = tpl.substitute({
-            'help' : g.help,
+            'kwargs' : json.dumps(g.kwargs,sort_keys=True, indent=4),
+            'help'   : g.help,
         })
-
-        print(Fore.LIGHTBLUE_EX + header)
-        print("="*len(header))
-        print(Fore.LIGHTBLUE_EX + content + Style.RESET_ALL)
-
+        page = "{}\n{}".format(header,content)
+        print(self.color_gen)
+        pydoc.pager(page)
+        
     def do_setGenerator(self, generator):
         """setGenerator [generator]
         sets active generator. if no arg: shows details about active generator"
         """
-        print(Fore.BLUE)
+        print(self.color_gen)
         if generator:
             self.egw.generator = generator
         self.do_detailGenerator(generator)
@@ -106,6 +138,55 @@ class EGWShell(cmd.Cmd):
         sets active generator. if no arg: shows details about active generator"
         """
         self.do_setGenerator(generator)
+
+    def do_paramsGenerator(self, arg):
+        """paramsGenerator [parameters]
+        edit parameters (kwargs) of active generator.
+        equivalent to '--params-generator=' on console.
+
+        optional args:
+            parameters (string) : generator's parameters in json-format
+
+        Example:
+            paramsGenerathor {"seed":1372403442, "colors":["#ff0000", "#00ff00", "#0000ff"]}' 
+        """
+        params = self.egw.generator.kwargs
+        if not arg:
+            params = json.dumps(params)
+        else:
+            params = arg
+        params = json.dumps(json.loads(params), indent=4, sort_keys=True) # prettify
+        status_msg = colored("ok.", "green")
+        finished = False
+        while not finished:
+            params = editor(params)
+            try:
+                self.egw.generator.kwargs = json.loads(params)
+            except:
+                print(colored("Ouch! Could not parse parameters...", "green"))
+                finished = not confirm_prompt("Wanna re-edit?")
+                print(finished)
+            else:
+                print(colored("ok.", "green"))
+                finished = True
+
+    def do_editGenerator(self, arg):
+        """alias of paramsGenerator"""
+        self.do_paramsGenerator(arg)
+
+    # ...
+
+    def do_run(self, arg):
+        """
+        run (egw-)job
+        """
+        print("TODO")
+
+    def do_printCLI(self, arg):
+        """
+        generate suiting egw command (handy for copy'n'paste)
+        """
+        print("TODO")
 
 def main(kwargs):
     init()